% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTCSAM02Params.FisheryCs.R
\name{calcTCSAM02Params.FisheryCs}
\alias{calcTCSAM02Params.FisheryCs}
\title{Calculate fishery catchability parameters for TCSAM02 based on values from a TCSAM2013 model.}
\usage{
calcTCSAM02Params.FisheryCs(obj, tpl, verbose = FALSE)
}
\arguments{
\item{obj}{\itemize{
\item tcsam2013.resLst object or tcsam2103.prs object
}}

\item{tpl}{\itemize{
\item character vector with file template
}}

\item{verbose}{\itemize{
\item flag to print debugging info
}}
}
\value{
file template with fishery catchability parameter values filled in
}
\description{
Function to calculate fishery catchability parameter values for TCSAM02 based on values from a TCSAM2013 model.
}
\details{
Calculates and returns a list of survey Q parameter values for a TCSAM02 model equivalent to those from a TCSAM2013 model run.
}
