% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTCSAM02Params.SelFcns.R
\name{calcTCSAM02Params.SelFcns}
\alias{calcTCSAM02Params.SelFcns}
\title{Calculate selectivity function parameters for TCSAM02 based on natural mortality rates from TCSAM2013.}
\usage{
calcTCSAM02Params.SelFcns(obj, tpl, verbose = FALSE)
}
\arguments{
\item{obj}{\itemize{
\item tcsam2013.rep object or tcsam2103.prs object
}}

\item{tpl}{\itemize{
\item character vector with file template
}}

\item{verbose}{\itemize{
\item flag to print debugging info
}}
}
\value{
file template with sel function values filled in
}
\description{
Function to calculate selectivity function parameter values from a TCSAM2013 model for TCSAM02.
}
\details{
Calculates and returns a list of selectivity function parameter values for a TCSAM02 model equivalent to those from a TCSAM2013 model run.
}
