% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTCSAM02Params.SurveyQs.R
\name{calcTCSAM02Params.SurveyQs}
\alias{calcTCSAM02Params.SurveyQs}
\title{Calculate survey Q parameters for TCSAM02 based on values from a TCSAM2013 model.}
\usage{
calcTCSAM02Params.SurveyQs(obj, tpl, verbose = FALSE)
}
\arguments{
\item{obj}{- tcsam2013.resLst object or tcsam2103.prs object}

\item{tpl}{- character vector with file template}

\item{verbose}{- flag to print debugging info}
}
\value{
file template with survey Q parameter values filled in
}
\description{
Function to calculate survey Q parameter values for TCSAM02 based on values from a TCSAM2013 model.
}
\details{
Calculates and returns a list of survey Q parameter values for a TCSAM02 model equivalent to those from a TCSAM2013 model run.
}

