% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createResLst.R
\name{createResLst}
\alias{createResLst}
\title{Create a tcsam02.rep object from components}
\usage{
createResLst(rep = NULL, prs = NULL, std = NULL, ofc = NULL)
}
\arguments{
\item{rep}{- tcsam02 rep object}

\item{prs}{- tcsam02 prs object}

\item{std}{- tcsam02 std object}

\item{ofc}{- tcsam02 ofc object}
}
\value{
a tcsam02.resLst object
}
\description{
Function to create a tcsam02.rep object from components.
}
\details{
None.
}
