% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelFits.R
\name{findOptDiffByCat}
\alias{findOptDiffByCat}
\title{Find differences between 'optimal' fits to models}
\usage{
findOptDiffByCat(o, n, str = "", type = "objfun", diffcrit = 10, print = FALSE)
}
\arguments{
\item{o}{\itemize{
\item category object from original model.fit object
}}

\item{n}{\itemize{
\item category object from model.fit object for comparison
}}

\item{str}{\itemize{
\item character string
}}

\item{type}{\itemize{
\item value for comparisons: 'objfun' or 'nll'
}}

\item{print}{\itemize{
\item flag to print 'critical' differences
}}
}
\value{
data.frame with elements:
\itemize{
\item{diff - difference in values}
\item{str - string indicating category, value, and difference}
}
}
\description{
Function to find differences between 'optimal' fits to models
}
\details{
Example: \cr
o: model.fits list object element (e.g., an element from the list from source('ModelFits.init.R'))\cr
n: corresponding  model.fits list object element for comparison\cr
#--loop over model fit categories and compare differences in 'objfun'\cr
for (ctg in names(o)) \cr
findOptDiffByCat(o[\link{ctg}],n[\link{ctg}],ctg,type="objfun",diffcrit=10);\cr
}
