% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelFits.R
\name{findOptDiffs}
\alias{findOptDiffs}
\title{Find differences between 'optimal' fits to models across all categories}
\usage{
findOptDiffs(o, n, type = "objfun", diffcrit = 10, showTop = 10)
}
\arguments{
\item{o}{\itemize{
\item tcsam02 model.fit list object
}}

\item{n}{\itemize{
\item tcsam02 model.fit list object for comparison
}}

\item{type}{\itemize{
\item value for comparisons: 'objfun' or 'nll'
}}

\item{diffcrit}{\itemize{
\item difference to consider critical
}}

\item{showTop}{\itemize{
\item max number of top differences to print
}}
}
\description{
Function to find differences between 'optimal' fits to models across all categories
}
