% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorr.R
\name{getCorr}
\alias{getCorr}
\title{Find substantially-correlated model parameters, based on the model "cor" file}
\usage{
getCorr(dfr = NULL, folder = NULL, model = "tcsam02",
  minLevel = 0.95)
}
\arguments{
\item{dfr}{- dataframe from a call to readCorFile, or NULL}

\item{folder}{- folder containing the model run (if dfr is NULL)}

\item{model}{- model name (default is "tcsam02")}

\item{minLevel}{- minimum correlation level to use to report parameters}
}
\value{
a dataframe
}
\description{
Function to find substantially-correlated model parameters, based on the model "cor" file.
}
\details{
This function extracts "substantially"-correlated parameters (ultimately) from the model
"cor" file, where "substantial" is based on the value of minLevel.
A dataframe with the parameter names and correlation values is returned.
}
