% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.AllScores.Abundance.R
\name{getMDFR.AllScores.Abundance1}
\alias{getMDFR.AllScores.Abundance1}
\title{Get all scores related to model fits to abundance time series for fleet data components}
\usage{
getMDFR.AllScores.Abundance1(rep, fleet.type, catch.type, verbose)
}
\arguments{
\item{rep}{\itemize{
\item single model report list object
}}

\item{fleet.type}{\itemize{
\item fleet type ('fishery' or 'survey')
}}

\item{catch.type}{\itemize{
\item catch type ('index','retained','discarded',or 'total')
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\value{
dataframe (not in canonical format)
}
\description{
Function to get all scores related to model fits to abundance time series for fleet data components.
}
\details{
Uses \code{\link[=getMDFR.ZScoresForABData]{getMDFR.ZScoresForABData()}}.
Returned dataframe is in canonical format.

This function is not exported. Use \code{\link[=getMDFR.AllScores.Abundance]{getMDFR.AllScores.Abundance()}}.

Column \code{type} indicates the type
of value, which is one of "z-score","observed","predicted","sdobs","xcv","stdv", and "useFlg".
}
