% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.AllScores.Biomass.R
\name{getMDFR.AllScores.Biomass}
\alias{getMDFR.AllScores.Biomass}
\title{Get all scores related to model fits to biomass time series for fleet data components}
\usage{
getMDFR.AllScores.Biomass(
  objs,
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "discarded", "total"),
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item single model report list object, or named list of them
}}

\item{fleet.type}{\itemize{
\item fleet type ('fishery' or 'survey')
}}

\item{catch.type}{\itemize{
\item catch type ('index','retained','discarded',or 'total')
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\value{
dataframe
}
\description{
Function to get model fits to biomass time series as z-scores for fleet data components.
}
\details{
Uses \code{\link[=getMDFR.AllScores.Biomass1]{getMDFR.AllScores.Biomass1()}} to get scores via \code{\link[=getMDFR.AllScoresForABData]{getMDFR.AllScoresForABData()}}.
Returned dataframe is in canonical format.

Column \code{type} indicates the type
of value, which is one of "z-score","observed","predicted","sdobs","xcv","stdv", and "useFlg".
}
