% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.CanonicalFormat.R
\name{getMDFR.CanonicalFormat}
\alias{getMDFR.CanonicalFormat}
\title{Convert dataframe to canonical format}
\usage{
getMDFR.CanonicalFormat(mdfr)
}
\arguments{
\item{mdfr}{\itemize{
\item dataframe to convert
}}
}
\value{
dataframe in canonical format, or NULL if input was not a dataframe
}
\description{
Function to convert dataframe to canonical format.
}
\details{
returns a dataframe in canonical format
\itemize{
\item{case - model case name (scenario)}
\item{process - 'population','fisheries',or 'surveys'}
\item{fleet - fishery or survey "fleet" name}
\item{category - data category (e.g., "index", "captured", etc.)}
\item{type - 'observed' or 'predicted',or ''}
\item{pc - parameter combination index, or ''}
\item{y - year or 'all'}
\item{x - sex or 'all'}
\item{m - maturity state or 'all'}
\item{s - shell condition or 'all'}
\item{z - size or 'all'}
\item{zp - size, \link{only for growth increments and growth transition matrices}}
\item{val - value}
\item{lci - lower ci}
\item{uci - upper ci}
}
}
