% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Data.ChelaHeightData.R
\name{getMDFR.Data.ChelaHeightData}
\alias{getMDFR.Data.ChelaHeightData}
\title{Get chela height data as a dataframe}
\usage{
getMDFR.Data.ChelaHeightData(objs, verbose = FALSE)
}
\arguments{
\item{objs}{\itemize{
\item single model report list object, or named list of them
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\value{
dataframe (or NULL)
}
\description{
Function to get chela height data as a dataframe.
}
\details{
Returned dataframe is in canonical format
}
