% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Data.EffortData.R
\name{getMDFR.Data.EffortData}
\alias{getMDFR.Data.EffortData}
\title{Get model fits to biomass for fleet data components}
\usage{
getMDFR.Data.EffortData(objs, verbose = FALSE)
}
\arguments{
\item{objs}{- single model resLst object, or named list of them}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
dataframe
}
\description{
Function to get effort data for multiple model scenarios
}
\details{
Returned dataframe is in canonical format
}
