% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Data.FleetSizeComps.R
\name{getMDFR.Data.FleetSizeComps}
\alias{getMDFR.Data.FleetSizeComps}
\title{Get fleet size comps by model scenario}
\usage{
getMDFR.Data.FleetSizeComps(
  objs,
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "total", "discard", "retained"),
  verbose = FALSE
)
}
\arguments{
\item{objs}{- single model report list object, or named list of them}

\item{fleet.type}{- fleet type ('fishery' or 'survey')}

\item{catch.type}{- catch type ('index','total','discard', or 'retained')}

\item{verbose}{- flag (T/F) to print diagnostic info}

\item{ci}{- confidence intervals for time series observations}
}
\value{
dataframe in canonical format
}
\description{
Function to get fleet size comps by model scenario.
}
\details{
Returned dataframe is in canonical format. "category" will be "n.at.z"
to indicate size compositions (to be consistent across the package).
}
