% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Data.GrowthData.R
\name{getMDFR.Data.GrowthData}
\alias{getMDFR.Data.GrowthData}
\title{Get growth data}
\usage{
getMDFR.Data.GrowthData(objs, verbose = FALSE)
}
\arguments{
\item{objs}{- single model report list object, or named list of them}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
dataframe (or NULL)
}
\description{
Function to get growth data.
}
\details{
Returned dataframe is in canonical format
}
