% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Data.MaturityOgiveData.R
\name{getMDFR.Data.MaturityOgiveData}
\alias{getMDFR.Data.MaturityOgiveData}
\title{Get maturity ogive data as a dataframe}
\usage{
getMDFR.Data.MaturityOgiveData(objs, verbose = FALSE)
}
\arguments{
\item{objs}{- single model report list object, or named list of them}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
dataframe (or NULL)
}
\description{
Function to get maturity ogive data as a dataframe.
}
\details{
Returned dataframe is in canonical format
}
