% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Data.TimeFrames.R
\name{getMDFR.Data.TimeFrames}
\alias{getMDFR.Data.TimeFrames}
\title{Get time frames for model data components as a dataframe}
\usage{
getMDFR.Data.TimeFrames(objs, verbose = FALSE)
}
\arguments{
\item{objs}{\itemize{
\item single model resLst object, or named list of them
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\value{
dataframe
}
\description{
Function to get time frames for model data components as a dataframe.
}
\details{
TODO: COMPLETE THIS!!
Returned dataframe has the following columns:
}
