% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.EffectiveNsForSizeComps.R
\name{getMDFR.EffectiveNsForSizeComps}
\alias{getMDFR.EffectiveNsForSizeComps}
\title{Get effective Ns for size comps as a dataframe}
\usage{
getMDFR.EffectiveNsForSizeComps(fits, mc, verbose = FALSE)
}
\arguments{
\item{fits}{\itemize{
\item list of fits for size compositions
}}

\item{mc}{\itemize{
\item model configuration list
}}

\item{verbose}{\itemize{
\item flag (T/F) to print dagnostic info
}}
}
\value{
dataframe
}
\description{
Get get effective Ns for size comps as a dataframe.
}
\details{
Uses \code{reshape2::melt(...)}.
Returned dataframe has columns:
\itemize{
\item{x - sex}
\item{m - maturity}
\item{s - shell condition}
\item{y - year}
\item{val - value}
\item{var - variable type ('input ss' or 'effective N')}
}
}
