% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Fisheries.CatchBiomass.R
\name{getMDFR.Fisheries.CatchBiomass}
\alias{getMDFR.Fisheries.CatchBiomass}
\title{Get fishery catch biomass time series from model results from TCSAM02 model runs as a dataframe}
\usage{
getMDFR.Fisheries.CatchBiomass(tcsams, category = c("captured", "discarded",
  "retained", "discard mortality", "index"), cast = "x", verbose = FALSE)
}
\arguments{
\item{tcsams}{- single tcsam02.rep object, tcsam02.resLst object, or named list of the latter}

\item{category}{- 'captured','discarded','retained', 'discard mortality', or 'index'}

\item{cast}{- casting formula (or NULL) for excluding x,m,s,z factor levels from an average-at-size across unspecified factors}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format
}
\description{
Function to get fishery catch biomass time series from model results from TCSAM02 model runs as a dataframe.
}
\details{
If 'cast' is not NULL (nor blank), the final cast'ing formula is
"case+process+fleet+category+type+pc+y+cast~.".
}

