% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Fisheries.RetFcns.R
\name{getMDFR.Fisheries.RetFcns}
\alias{getMDFR.Fisheries.RetFcns}
\title{Get fishery retention functions from model results from TCSAM02 model runs as a dataframe}
\usage{
getMDFR.Fisheries.RetFcns(tcsams, verbose = FALSE, cast = "y+x")
}
\arguments{
\item{tcsams}{- single TCSAM02 object, TCSAM02.resLst object, or named list of the latter}

\item{verbose}{- flag (T/F) to print debug info}

\item{cast}{- casting formula for excluding y,x,m,s factor levels from an average-at-size across unspecified factors}
}
\value{
dataframe in canonical format
}
\description{
Function to get fishery retention functions from model results from TCSAM02 model runs as a dataframe.
}
\details{
Extracts fishery retention functions.
}
