% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Fits.EffectiveNs.R
\name{getMDFR.Fits.EffectiveNs}
\alias{getMDFR.Fits.EffectiveNs}
\title{Get effective N's for fleet size compositions}
\usage{
getMDFR.Fits.EffectiveNs(
  objs,
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "total", "discard", "retained"),
  verbose = FALSE
)
}
\arguments{
\item{objs}{- single model report list object, or named list of them}

\item{fleet.type}{- fleet type ('fishery' or 'survey')}

\item{catch.type}{- cath type ('index','total','discard', or 'retained')}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
dataframe
}
\description{
Function to get effective N's for fleet size compositions.
}
\details{
Returned dataframe is in canonical format
}
