% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Fits.EffortData.R
\name{getMDFR.Fits.EffortData}
\alias{getMDFR.Fits.EffortData}
\title{Get model fits to effort data}
\usage{
getMDFR.Fits.EffortData(objs, verbose = FALSE)
}
\arguments{
\item{objs}{\itemize{
\item single tcsam02 resLst object, or named list of them
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\value{
dataframe
}
\description{
Function to get model fits to effort data.
}
\details{
Returned dataframe is in canonical format
}
