% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Fits.FleetData.R
\name{getMDFR.Fits.FleetData}
\alias{getMDFR.Fits.FleetData}
\title{Get model fits to biomass for fleet data components}
\usage{
getMDFR.Fits.FleetData(
  objs,
  fleet.type = c("survey", "fishery"),
  data.type = c("abundance", "biomass", "n.at.z"),
  catch.type = c("index", "total", "discard", "retained"),
  ci = 0.8,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item single model report list object, or named list of them
}}

\item{fleet.type}{\itemize{
\item fleet type ('fishery' or 'survey')
}}

\item{data.type}{\itemize{
\item data type ('abundance', 'biomass', or 'n.at.z')
}}

\item{catch.type}{\itemize{
\item catch type ('index','total','discard', or 'retained')
}}

\item{ci}{\itemize{
\item confidence intervals for observations
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\value{
dataframe
}
\description{
Function to get model fits to biomass data for fleet data components.
}
\details{
Uses \code{\link[=getMDFR.FitsForABData]{getMDFR.FitsForABData()}} and \code{\link[=getMDFR.FitsForSizeComps]{getMDFR.FitsForSizeComps()}}.
Returned dataframe is in canonical format
}
