% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Fits.GrowthData.R
\name{getMDFR.Fits.GrowthData}
\alias{getMDFR.Fits.GrowthData}
\title{Get model fits to growth data}
\usage{
getMDFR.Fits.GrowthData(objs, verbose = FALSE)
}
\arguments{
\item{objs}{- single model report list object, or named list of them}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
dataframe
}
\description{
Function to get model fits to growth data.
}
\details{
Returned dataframe is in canonical format
}

