% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.FitsForABData.R
\name{getMDFR.FitsForABData}
\alias{getMDFR.FitsForABData}
\title{Get fits to abundance or biomass data as a dataframe}
\usage{
getMDFR.FitsForABData(afits, ci = 0.95, verbose = FALSE)
}
\arguments{
\item{afits}{\itemize{
\item 
}}

\item{ci}{\itemize{
\item confidence intervals
}}

\item{verbose}{\itemize{
\item flag (T/F) to print dagnostic info
}}
}
\value{
dataframe (not in "canonical" format)
}
\description{
Function to get fits to abundance or biomass data as a dataframe.
}
\details{
Fits that have both observations and model estimates identically zero are excluded
from the returned dataframe. Note that the returned dataframe is NOT in canonical format.
}
