% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.ManagementQuantities.R
\name{getMDFR.ManagementQuantities}
\alias{getMDFR.ManagementQuantities}
\title{Get management quantities from multiple model scenarios}
\usage{
getMDFR.ManagementQuantities(objs, verbose = FALSE)
}
\arguments{
\item{objs}{\itemize{
\item single model resLst object, or named list of them
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\value{
dataframe
}
\description{
Function to get management quantities from multiple model scenarios
}
\details{
Returned dataframe is in canonical format
}
