% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.OFCs.NonDataComponents.R
\name{getMDFR.OFCs.FDevPenalties}
\alias{getMDFR.OFCs.FDevPenalties}
\title{Get F-devs penalties from the objective function for a TCSAM02 model as a dataframe}
\usage{
getMDFR.OFCs.FDevPenalties(fits, verbose = FALSE)
}
\arguments{
\item{fits}{- model.fits list element from a tcsam02.rep object}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
dataframe
}
\description{
Function to get F-devs penalties  from the objective function for a TCSAM02 model as a dataframe.
}
\details{
Called by \code{getMDFR.OFCs.NonDataComponents()}. Returned dataframe has columns
\itemize{
  \item{case - model case [blank, to be filled in by calling function]}
  \item{category - objective function component category}
  \item{type - type name}
  \item{element - element name}
  \item{level - level name}
  \item{wgt - likelihood weight}
  \item{nll - negative log-likelihood (unweighted)}
  \item{objfun - objective function value}
}
}
