% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.OFCs.FleetData.R
\name{getMDFR.OFCs.FleetData}
\alias{getMDFR.OFCs.FleetData}
\title{Get objective function components for growth data from a tcsam02.resLst or tcsam02.rep object}
\usage{
getMDFR.OFCs.FleetData(obj, category = c("surveys", "fisheries"),
  verbose = FALSE)
}
\arguments{
\item{obj}{- a tcsam02.resLst or tcsam02.rep object}

\item{category}{- fleet category ("surveys" or "fisheries")}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
a dataframe
}
\description{
Function to get objective function components for growth data from a tcsam02.resLst or tcsam02.rep object.
}
\details{
Returned dataframe has columns:
\itemize{
  \item{case - model case (blank, to be filled in by caller)}
  \item{category - fleet category}
  \item{fleet - fleet name}
  \item{catch.type - catch type}
  \item{data.type - data type}
  \item{fit.type - fit type}
  \item{nll.type - likelihood type}
  \item{y - year}
  \item{x - sex}
  \item{m - maturity state}
  \item{s  - shell condition}
  \item{ rmse - root mean squared error}
  \item{ wgt - likelihood weight}
  \item{nll - (unweighted) negative log-likelihood}
  \item{objfun - objective function value}
}
}
