% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.OFCs.MaturityData.R
\name{getMDFR.OFCs.MaturityData}
\alias{getMDFR.OFCs.MaturityData}
\title{Get objective function components for maturity data from a tcsam02.rep object}
\usage{
getMDFR.OFCs.MaturityData(obj, verbose = FALSE)
}
\arguments{
\item{obj}{- a tcsam02.resLst or tcsam02.rep object}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
a dataframe
}
\description{
Function to get objective function components for maturity data from a tcsam02.rep object.
}
\details{
Returned dataframe has columns:
\itemize{
  \item{case - model case (blank, to be filled in by caller)}
  \item{category - "maturity data"}
  \item{fleet - dummy value ("")}
  \item{catch.type - dummy value ("")}
  \item{data.type - data type ("maturity data")}
  \item{fit.type - dummy value ("")}
  \item{nll.type - likelihood type}
  \item{y - year}
  \item{x - sex}
  \item{m - maturity state ("")}
  \item{s - shell condition ("new shell")}
  \item{ rmse - root mean squared error}
  \item{ wgt - likelihood weight}
  \item{nll - (unweighted) negative log-likelihood}
  \item{objfun - objective function value}
}
}
