% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.OFCs.NonDataComponents.R
\name{getMDFR.OFCs.NonDataComponents}
\alias{getMDFR.OFCs.NonDataComponents}
\title{Get non-data components from the objective function for a TCSAM02 model as a dataframe}
\usage{
getMDFR.OFCs.NonDataComponents(obj, categories = c("penalties", "penFDevs",
  "components", "priors"), verbose = FALSE)
}
\arguments{
\item{obj}{- a tcsam02.resLst or tcsam02.rep object}

\item{categories}{- non-data-related objective function components to get ("all","penalties","penFDevs","components","priors")}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
dataframe
}
\description{
Function to get non-data components from the objective function for a TCSAM02 model as a dataframe.
}
\details{
Returned dataframe has columns
\itemize{
  \item{case - model case [blank, to be filled in by calling function]}
  \item{category - objective function component category}
  \item{type - type name}
  \item{element - element name}
  \item{level - level name}
  \item{wgt - likelihood weight}
  \item{nll - negative log-likelihood (unweighted)}
  \item{objfun - objective function value}
}
}
