% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.CohortProgression.R
\name{getMDFR.Pop.CohortProgression}
\alias{getMDFR.Pop.CohortProgression}
\title{Get cohort progression time series from model results from TCSAM02 model runs as a dataframe}
\usage{
getMDFR.Pop.CohortProgression(tcsams, cast = "x+m+s+z", verbose = FALSE)
}
\arguments{
\item{tcsams}{- single tcsam02.rep object, tcsam02.resLst object, or named list of the latter}

\item{cast}{- casting formula for excluding x,m,s,z factor levels from a sum across the unspecified factors}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format
}
\description{
Function to get cohort progression time series from model results from TCSAM02 model runs as a dataframe.
}
\details{
Extracts the estimated cohort abundance time series.
}
