% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.GrowthMatrices.R
\name{getMDFR.Pop.GrowthMatrices}
\alias{getMDFR.Pop.GrowthMatrices}
\title{Get growth transition matrices from model results from TCSAM02 model runs as a dataframe}
\usage{
getMDFR.Pop.GrowthMatrices(tcsams, path = "mp/T_list/T_czz", verbose = FALSE)
}
\arguments{
\item{tcsams}{- single tcsam02.rep, tcsam02.resLst, named list of the latter, or general list}

\item{path}{- path into tcsams object at which to find the transition arrays}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format.
Note that 'z' is pre-molt size, 'zp' is post-molt size.
}
\description{
Function to get growth transition matrices from model results from TCSAM02 model runs as a dataframe.
}
\details{
Extracts growth transition matrices.
}
