% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.GrowthTransitionMatrices.R
\name{getMDFR.Pop.GrowthTransitionMatrices}
\alias{getMDFR.Pop.GrowthTransitionMatrices}
\title{Get growth transition matrices from model results from TCSAM02 model runs as a dataframe}
\usage{
getMDFR.Pop.GrowthTransitionMatrices(tcsams, verbose = FALSE)
}
\arguments{
\item{tcsams}{- single tcsam02.rep, tcsam02.resLst, or named list of the latter}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format.
Note that 'z' is pre-molt size, 'zp' is post-molt size.
}
\description{
Function to get growth transition matrices from model results from TCSAM02 model runs as a dataframe.
}
\details{
Extracts growth transition matrices.
}

