% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.MaturityOgives.R
\name{getMDFR.Pop.MaturityOgives}
\alias{getMDFR.Pop.MaturityOgives}
\title{Get population maturity ogives by year from model results from TCSAM02 model runs as a dataframe}
\usage{
getMDFR.Pop.MaturityOgives(tcsams, verbose = FALSE)
}
\arguments{
\item{tcsams}{\itemize{
\item single tcsam02.rep object, tcsam02.resLst object, or named list of the latter
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debug info
}}
}
\value{
dataframe in canonical format
}
\description{
Function to get population maturity ogives by year from model results from TCSAM02 model runs as a dataframe.
}
\details{
Extracts the estimated population maturity ogives by year.
}
