% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.RecSizeDistribution.R
\name{getMDFR.Pop.RecSizeDistribution}
\alias{getMDFR.Pop.RecSizeDistribution}
\title{Get recruitment size distribution from model results from TCSAM02 model runs as a dataframe}
\usage{
getMDFR.Pop.RecSizeDistribution(tcsams, verbose = FALSE)
}
\arguments{
\item{tcsams}{\itemize{
\item single tcsam02.rep object, tcsam02.resLst object, or named list of the latter
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debug info
}}
}
\value{
dataframe in canonical format.
}
\description{
Function to get recruitment size distribution from model results from TCSAM02 model runs as a dataframe.
}
\details{
Extracts recruitment size distributions (values are \emph{not} scaled by bin size).
}
