% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.WatZ.R
\name{getMDFR.Pop.WatZ}
\alias{getMDFR.Pop.WatZ}
\title{Get individual weights-at-size from TCSAM02 model runs as a dataframe}
\usage{
getMDFR.Pop.WatZ(tcsams, verbose = FALSE)
}
\arguments{
\item{tcsams}{\itemize{
\item single tcsam02.rep object, tcsam02.resLst object, or named list of the latter
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debug info
}}
}
\value{
dataframe in canonical format
}
\description{
Function to get individual weights-at-size from TCSAM02 model runs as a dataframe.
}
\details{
Extracts the individual weights-at-size by sex, maturity state, and size. Units are in kg.
}
