% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcns_Projections.R
\name{getMDFR.Projections}
\alias{getMDFR.Projections}
\title{Get a TCSAM02 projections list (a tcsam02.projList object) as a \link[tibble:tibble]{tibble::tibble}}
\usage{
getMDFR.Projections(projList)
}
\arguments{
\item{projList}{\itemize{
\item projections list parsed using \link{getProjectionsList(projFile)}
}}
}
\value{
a \link[tibble:tibble]{tibble::tibble} representing a TCSAM02 projections list, or NULL. The returned object will be of class 'tcsam02.projDFR'.
}
\description{
Function to get a TCSAM02 projections list (a tcsam02.projList object) as a \link[tibble:tibble]{tibble::tibble}.
}
\details{
The returned object will be a \link[tibble:tibble]{tibble::tibble} of class 'tcsam02.projDFR'.
}
