% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.R
\name{getMDFR}
\alias{getMDFR}
\title{Get model arrays as a melted dataframe from TCSAM02 model runs}
\usage{
getMDFR(path, tcsams, verbose = FALSE)
}
\arguments{
\item{path}{\itemize{
\item path in models to requested array (using '/' as separator for list levels)
}}

\item{tcsams}{\itemize{
\item single tcsam02.resLst object, or named list of such
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostics
}}
}
\value{
Melted dataframe (ala package reshape2).
}
\description{
Function to get model objects as a melted dataframe from TCSAM2015 model runs.
}
\details{
Returned dataframe is a melted (ala reshape2) version of the requested array,
with additional columns 'model' and 'modeltype' appended at the "right". The array value
is in column 'val'. Uses \code{reshape2::melt(...)}.
}
