% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.SdRep.DerivedVars.R
\name{getMDFR.SdRep.DerivedVars}
\alias{getMDFR.SdRep.DerivedVars}
\title{Get sd_report info on derived quantities as a dataframe}
\usage{
getMDFR.SdRep.DerivedVars(tcsams, ci = 0.95, verbose = FALSE)
}
\arguments{
\item{tcsams}{\itemize{
\item a tcsam02.resLst object or named list of such
}}

\item{ci}{\itemize{
\item confidence interval
}}

\item{verbose}{\itemize{
\item flag to print debugging info
}}
}
\value{
a dataframe
}
\description{
Function to get sd_report info on derived time series estimates as a dataframe.
}
\details{
Derived variables are
"sdrAvgRec","sdrBmsy","sdrFmsy","sdrMSY","sdrCurB","sdrPrjB","sdrOFL"
}
