% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Surveys.Availability.R
\name{getMDFR.Surveys.Availability}
\alias{getMDFR.Surveys.Availability}
\title{Get annual survey availabilities from model results from TCSAM02 model runs as a dataframe}
\usage{
getMDFR.Surveys.Availability(tcsams, cast = "x", verbose = FALSE)
}
\arguments{
\item{tcsams}{- single tcsam02.rep object, tcsam02.resLst object, or named list of the latter}

\item{cast}{- casting formula for excluding x,m,s factor levels from an average across unspecified factors}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format
}
\description{
Function to get annual survey availabilities from TCSAM02 model runs as a dataframe.
}
\details{
Extracts annual survey availabilities.
}
