% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Surveys.CaptureProbs.R
\name{getMDFR.Surveys.CaptureProbs}
\alias{getMDFR.Surveys.CaptureProbs}
\title{Get survey capture probability (Q_vyxmsz) functions from TCSAM02 model runs as a dataframe}
\usage{
getMDFR.Surveys.CaptureProbs(tcsams, verbose = FALSE, cast = "y+x")
}
\arguments{
\item{tcsams}{\itemize{
\item single tcsam02.rep object, tcsam02.resLst object, or named list of the latter
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debug info
}}

\item{cast}{\itemize{
\item casting formula for excluding y,x,m,s factor levels from an average-at-size across unspecified factors
}}
}
\value{
dataframe in canonical format
}
\description{
Function to get survey scapture probability (Q_vyxmsz) functions from TCSAM02 model runs as a dataframe.
}
\details{
Extracts survey capture probability functions (Q_vyxmsz).
}
