% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.ZScores.Biomass.R
\name{getMDFR.ZScores.Biomass}
\alias{getMDFR.ZScores.Biomass}
\title{Get model fits to biomass time series as z-scores for fleet data components}
\usage{
getMDFR.ZScores.Biomass(
  objs,
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "discarded", "total"),
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item single model report list object, or named list of them
}}

\item{fleet.type}{\itemize{
\item fleet type ('fishery' or 'survey')
}}

\item{catch.type}{\itemize{
\item catch type ('index','retained','discarded',or 'total')
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\value{
dataframe
}
\description{
Function to get model fits to biomass time series as z-scores for fleet data components.
}
\details{
Uses \code{\link[=getMDFR.ZScoresForABData]{getMDFR.ZScoresForABData()}}.
Returned dataframe is in canonical format.
}
