% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.ZScores.Biomass.R
\name{getMDFR.ZScores.Biomass1}
\alias{getMDFR.ZScores.Biomass1}
\title{Get model fits to biomass time series as z-scores for fleet data components}
\usage{
getMDFR.ZScores.Biomass1(rep, fleet.type, catch.type, verbose)
}
\arguments{
\item{rep}{- single model report list object}

\item{fleet.type}{- fleet type ('fishery' or 'survey')}

\item{catch.type}{- catch type ('index','retained','discarded',or 'total')}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
dataframe
}
\description{
Function to get model fits to biomass time series as z-scores for fleet data components.
}
\details{
Uses \code{getMDFR.ZScoresForABData()}.
Returned dataframe is in canonical format.
}
