% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.ZScores.PrNatZ1.R
\name{getMDFR.ZScores.PrNatZ1}
\alias{getMDFR.ZScores.PrNatZ1}
\title{Get residuals from size frequency fits as dataframe}
\usage{
getMDFR.ZScores.PrNatZ1(
  fits,
  mc,
  type = c("pearsons", "nlls"),
  verbose = FALSE
)
}
\arguments{
\item{fits}{- list of fits}

\item{mc}{- model configuration list}

\item{type}{- type of residuals ('pearsons' or 'nlls')}

\item{verbose}{- flag (T/F) to print dagnostic info}
}
\value{
dataframe
}
\description{
Get Pearson's residuals and negative log-likelihood components
from size frequency fits as dataframe.
}
\details{
Uses \code{reshape2::melt(...)}.
Returned dataframe has columns:
\itemize{
 \item{x - sex}
 \item{m - maturity}
 \item{s - shell condition}
 \item{y - year}
 \item{z - size bin}
 \item{val - value}
 \item{var - variable type}
}
}
