% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.ZScoresForABData.R
\name{getMDFR.ZScoresForABData}
\alias{getMDFR.ZScoresForABData}
\title{Get z-scores and other model fits info for abundance, biomass data from fits to data as dataframe}
\usage{
getMDFR.ZScoresForABData(afits, verbose = FALSE)
}
\arguments{
\item{afits}{\itemize{
\item 
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\value{
dataframe
}
\description{
Function to get z-scores and other model fits info for abundance, biomass data from fits to data as dataframe.
}
\details{
The returned dataframe is \strong{NOT} in canonical format and has columns
x,m,s,y,val,var,process,fleet,category,case. Column \code{var} gives the type
of value, which is one of "z-score","observed","predicted","sdobs","xcv","stdv", and "useFlg".
}
