% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOFCs.R
\name{getOFCs}
\alias{getOFCs}
\title{Get model components in the objective function as a tcsam02.ofc (melted dataframe) object}
\usage{
getOFCs(repObjs, mdl = NULL, verbose = FALSE)
}
\arguments{
\item{repObjs}{- tcsam02 model report object or list of such}

\item{mdl}{- name to associate with model results object}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
a melted dataframe with class tcsam02.ofc
}
\description{
Function to get model components in the objective function as a tcsam02.ofc (melted dataframe) object.
}
\details{
If repObjs is a list of tcsam02 model report objects (tcsam02.rep), then the function
is called recursively for each object, with the associated list component name used as
mdl. If repObjs is a tcsam02 model report object and mdl is NULL (the default), then
repObjs$mc$configName is used as the model name.

The returned dataframe has columns named
"model", "type", "category", "name", "level", "variable", and "value".

The "variable" column indicates whether the "value" is a weight ('wgt'),
negative log-likelihood ('nll'), or objective function value ('objfun').
}

