% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcns_Projections.R
\name{getProjectionsList}
\alias{getProjectionsList}
\title{Get a TCSAM02 projections list by reading a TCSAM02 calcProjections.R file}
\usage{
getProjectionsList(fn)
}
\arguments{
\item{repFile}{\itemize{
\item projections file from a TCSAM02 model run to source (default="calcProjections.final.R")
}}

\item{verbose}{\itemize{
\item flag (T/F) to prnt diagnostic info
}}
}
\value{
a TCSAM02 projections list, or NULL. The returned object will be a list of class 'tcsam02.projList'.
}
\description{
Function to get a TCSAM02 projections list by reading a TCSAM02 calcProjections.R file.
}
\details{
The returned object will be a list of class 'tcsam02.projList'.
}
