% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResLst.R
\name{getResLst}
\alias{getResLst}
\title{Create a tcsam02.resLst object from a model run}
\usage{
getResLst(
  inp.dir = NULL,
  rep = "tcsam02.rep",
  model = "tcsam02",
  prsType = c("all", "active"),
  verbose = FALSE
)
}
\arguments{
\item{inp.dir}{- a folder with model output}

\item{rep}{- filename to read for rep object}

\item{model}{- name of model executable}

\item{prsType}{-  type ('all' or 'active') of parameters to read}

\item{verbose}{- flag (T/F) to prnt diagnostic info}
}
\value{
a tcsam02.resLst object.
}
\description{
Function to create a tcsam02.resLst object from a model run.
}
\details{
Uses \code{tcltk::tk_chose.dir} to open a file dialog to select model directory
if inp.dir is NULL. A tcsam02.resLst object is a list with elements
\itemize{
 \item{rep - a tcsam02.rep object, or NULL}
 \item{prs - a tcsam02.prs object, or NULL}
 \item{std - a tcsam02.std object, or NULL}
 \item{ofc - a tcsam02.ofc object, or NULL}
}
}
