% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResLsts.R
\name{getResLsts}
\alias{getResLsts}
\title{Create a list of tcsam02.resLst objects, from a set of model runs}
\usage{
getResLsts(
  inp.dirs = NULL,
  rep = "tcsam02.rep",
  model = "tcsam02",
  prsType = c("all", "active")
)
}
\arguments{
\item{inp.dirs}{\itemize{
\item a dataframe with columns "case" and "path"
}}

\item{rep}{\itemize{
\item filename to read for rep object
}}

\item{model}{\itemize{
\item name of model executable
}}

\item{prsType}{\itemize{
\item type ('all' or 'active') of parameters to read
}}
}
\value{
a list of tcsam02.resLst objects.
}
\description{
Function to create a list of tcsam02.resLst objects from a set of model runs.
}
\details{
Uses \code{getResLst} to open a folders with model runs listed in inp.dirs.
}
