% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStd.R
\name{getStd}
\alias{getStd}
\title{Get a TCSAM02 std object by reading a .std file}
\usage{
getStd(stdFile = NULL, verbose = FALSE)
}
\arguments{
\item{stdFile}{- std file from a TCSAM02 model run. can be NULL.}

\item{verbose}{- flag (T/F) to prnt diagnostic info}
}
\value{
std model object (a list). The returned object will be a list of class 'tcsam02.std'.
}
\description{
Function to get a TCSAM02 std object by reading a .std file
}
\details{
If \code{stdFile} is NULL, the user will be prompted to identify a
TCSAM02 model report file from which to source the results object.
The returned object will be a list of class 'tcsam02.std'.
}
