% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.ExtractOFLResults.R
\name{mcmc.ExtractOFLResults}
\alias{mcmc.ExtractOFLResults}
\title{Extract OFL results from an mcmc run to a dataframe}
\usage{
mcmc.ExtractOFLResults(mcmc)
}
\arguments{
\item{mcmc}{- mcmc list from TCSAM02 mdoel run}
}
\value{
dataframe with
}
\description{
Function to extract OFL results from an mcmc run to a dataframe
}
\details{
none.
}
