% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.ExtractVars.R
\name{mcmc.ExtractVar}
\alias{mcmc.ExtractVar}
\title{Extract a variable from an MCMC results list}
\usage{
mcmc.ExtractVar(mcmc, vr)
}
\arguments{
\item{mcmc}{\itemize{
\item an MCMC results list from a TCSAM02 model
}}

\item{vr}{\itemize{
\item variable name to extract
}}
}
\value{
a vector or an array, depending on the variable
}
\description{
Function to extract a variable from an MCMC results list.
}
\details{
handles numeric and array mcmc variables. If the variable is a
n-dimensional array, the returned array will be (n+1)-dimensional, with the
mcmc iteration as the left-most dimension.
}
