% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.ExtractVarsFromChain.R
\name{mcmc.ExtractVarFromChain}
\alias{mcmc.ExtractVarFromChain}
\title{Extract a variable from an MCMC results list for a single chain}
\usage{
mcmc.ExtractVarFromChain(mcmc, vr, verbose = FALSE)
}
\arguments{
\item{mcmc}{\itemize{
\item an MCMC results list for a single chain from a TCSAM02 model
}}

\item{vr}{\itemize{
\item variable name to extract
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
a list, vector, or array, depending on the variable
}
\description{
Function to extract a variable from an MCMC results list for a single chain.
}
\details{
handles lists of numeric and array mcmc variables. If the variable is a
n-dimensional array, the returned array will be (n+1)-dimensional, with the
mcmc iteration as the left-most dimension.
}
