% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.ExtractVars.R
\name{mcmc.ExtractVars}
\alias{mcmc.ExtractVars}
\title{Extract all variables from an MCMC results list}
\usage{
mcmc.ExtractVars(mcmc)
}
\arguments{
\item{mcmc}{- an MCMC results list from a TCSAM02 model}
}
\value{
a list, with elements named for each variable
}
\description{
Function to extract all variables from an MCMC results list.
}
\details{
Calls mcmc.ExtractVar to extract individual variables.
}
