% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.ExtractVarsFromChain.R
\name{mcmc.ExtractVarsFromChain}
\alias{mcmc.ExtractVarsFromChain}
\title{Extract all variables from an MCMC results list for a single chain}
\usage{
mcmc.ExtractVarsFromChain(mcmc, verbose = FALSE)
}
\arguments{
\item{mcmc}{\itemize{
\item an MCMC results list for a single chain from a TCSAM02 model
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
a list, with elements named for each variable
}
\description{
Function to extract all variables from an MCMC results list for a single chain.
}
\details{
Calls \code{\link[=mcmc.ExtractVarFromChain]{mcmc.ExtractVarFromChain()}} to extract individual variables.
}
