% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.getMDFR.Projections.R
\name{mcmc.getMDFR.Projections}
\alias{mcmc.getMDFR.Projections}
\title{Convert a single chain MCMC list of projection lists to a \link[tibble:tibble]{tibble::tibble}}
\usage{
mcmc.getMDFR.Projections(mcmcList)
}
\arguments{
\item{mcmcList}{\itemize{
\item a single chain MCMC list of projection lists parsed using \link{mcmc.ReadMMCMC(projFile)}
}}
}
\value{
a \link[tibble:tibble]{tibble::tibble} representing a single chain MCMC list of projection lists, or NULL. The returned object will be of class 'tcsam02.mcmc.projDFR'.
}
\description{
Function to convert a single chain MCMC list of projection lists to a \link[tibble:tibble]{tibble::tibble}.
}
\details{
The returned object will be a \link[tibble:tibble]{tibble::tibble} of class 'tcsam02.mcmc.projDFR'.
}
