% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.InvertList.R
\name{mcmc.InvertList}
\alias{mcmc.InvertList}
\title{Invert an mcmc list}
\usage{
mcmc.InvertList(mcmc, verbose = FALSE)
}
\arguments{
\item{mcmc}{- mcmc list}

\item{verbose}{- T/F to print diagnostic info}
}
\value{
a dataframe
}
\description{
Function to invert an mcmc list
}
\details{
None
}
