% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.InvertListElement.R
\name{mcmc.InvertListElement}
\alias{mcmc.InvertListElement}
\title{Invert an element of a sublist list}
\usage{
mcmc.InvertListElement(mcmc, element = "", verbose = FALSE)
}
\arguments{
\item{mcmc}{\itemize{
\item an mcmc results list from a TCSAM02 model run
}}

\item{element}{\itemize{
\item name of a list element to extract
}}

\item{verbose}{\itemize{
\item T/F to print diagnostic info
}}
}
\value{
a list
}
\description{
Function to invert an element of an mcmc list
}
\details{
None
}
