% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.InvertMCMCList.R
\name{mcmc.InvertMCMCList}
\alias{mcmc.InvertMCMCList}
\title{Invert an mcmc list}
\usage{
mcmc.InvertMCMCList(mcmc, verbose = FALSE)
}
\arguments{
\item{mcmc}{\itemize{
\item mcmc list
}}

\item{verbose}{\itemize{
\item T/F to print diagnostic info
}}
}
\value{
a dataframe
}
\description{
Function to invert an mcmc list
}
\details{
None
}
