% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.InvertListElement.R
\name{mcmc.InvertVectorElement}
\alias{mcmc.InvertVectorElement}
\title{Invert a list of mcmc results for a numeric vector element to a dataframe}
\usage{
mcmc.InvertVectorElement(mcmc, el, verbose = FALSE)
}
\arguments{
\item{mcmc}{\itemize{
\item list of mcmc results
}}

\item{el}{\itemize{
\item name of an element that is a numeric vector with length >1 to extract, as character string
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagositc info
}}
}
\value{
a dataframe
}
\description{
Function to invert a list of mcmc results for a numeric vector element to a dataframe.
}
\details{
Throws an error if "el" is not an element of the mcmc results list, if el is not numeric, and if
length(el) is not > 1.
}
