% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.ParseMCMC.ProjectionsFile.R
\name{mcmc.ParseMCMC.Projections}
\alias{mcmc.ParseMCMC.Projections}
\title{Parse a TCSAM02 calcProjections.mcmc.R file}
\usage{
mcmc.ParseMCMC.Projections(path, verbose = FALSE)
}
\arguments{
\item{path}{\itemize{
\item path to the calcProjections.mcmc.R file to parse
}}

\item{verbose}{\itemize{
\item flag to print extra information
}}
}
\value{
list of individual mcmc projection results lists
}
\description{
Function to parse a calcProjections.mcmc.R file and return a list of lists
of mcmc projection results.
}
\details{
Uses \code{\link[=mcmc.ReadMCMC]{mcmc.ReadMCMC()}} to read the file
}
