% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.PlotABC.R
\name{mcmc.PlotABC}
\alias{mcmc.PlotABC}
\title{Function to plot ABC and OFL calculations from mcmc.CalcABC(...)}
\usage{
mcmc.PlotABC(dfrABCs, dfrOFLs, xlims = NULL)
}
\arguments{
\item{xlims}{- range for x-axis in plot}

\item{dfrABC}{- dataframe with selected values for ABC.buff, ABC.pstar, and OFL by case}

\item{dfrOFL}{- empirical cdfs for OFL, by case}
}
\value{
ggplot object
}
\description{
This function plots ABC and OFL from TCSAM02 mcmc output.
}
\details{
Inputs dfrABCs and dfrOFLs are outputs from \code{mcmc.CalcABC}
}
