% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.ReadMCMC.R
\name{mcmc.ReadMCMC}
\alias{mcmc.ReadMCMC}
\title{Parse TCSAM02 MCMC.R file}
\usage{
mcmc.ReadMCMC(fn, verbose = FALSE)
}
\arguments{
\item{fn}{\itemize{
\item the filename of te MCMC.R file to parse
}}

\item{verbose}{\itemize{
\item flag to print extra information
}}
}
\value{
list of individual mcmc results lists
}
\description{
Function to parse a TCSAM02 MCMC.R file and return a list of lists
of mcmc results.
}
\details{
Concatenates text lines up to each ";", parses the resulting string and
evaluates the expression.
}
