% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDynamicB0.R
\name{plotDynamicB0}
\alias{plotDynamicB0}
\title{Plot dyanmic B0 results}
\usage{
plotDynamicB0(obj, B100 = NULL)
}
\arguments{
\item{obj}{\itemize{
\item tcsam.resLst, tcsam.rep, or list object
}}

\item{B100}{\itemize{
\item B100 from OFL calculation
}}
}
\value{
ggplot object
}
\description{
Fuction to plot dynamic B0 details.
}
\details{
Extracts the dynamic B0 results from a TCSAM02 model run and plots
time series of dynamic B0 and MMB for comparison.
}
