% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMQs.QuadPlot.R
\name{plotMQs.QuadPlot}
\alias{plotMQs.QuadPlot}
\title{Make a quad plot of F vs. B over time}
\usage{
plotMQs.QuadPlot(tcsams, dfrMQs, aYr = 2022, min_year = 1980)
}
\arguments{
\item{tcsams}{\itemize{
\item a named list of tcsam02.resLst objects
}}

\item{dfrMQs}{\itemize{
\item dataframe with Fmsy, Bmsy, Fofl and OFL by case
}}

\item{aYr}{\itemize{
\item assessment year (default: 2022)
}}

\item{min_year}{\itemize{
\item minimum year to include in plot (default: 1980)
}}
}
\value{
cowplot grid graphics object
}
\description{
Function to make a quad plot of F vs. B over time.
}
\details{
The standard quad plot is faceted by model case. One possibility for
obtaining a suitable dfrMQs is output from \code{\link[=sdRep.CalcABC]{sdRep.CalcABC()}}.
}
