% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMSE.R
\name{runMSE}
\alias{runMSE}
\title{Function to run the TCSAM02 MSE for one harvest strategy}
\usage{
runMSE(os = "osx", topLevelFolder = ".", model = "tcsam02",
  path2model = "", HCR = 1, numRuns = 3, minRunID = 1,
  firstYr = 2018, numYrs = 10, runBaseModel = TRUE,
  baseModelInfo = list(path = ".", configFile = "ModelConfig.inp",
  datasets = list(file = "Datasets.inp", components = list(Bio = "Info",
  Fishery = c("TCF", "SCF", "GTF", "RKF"), Survey = "NMFS", Growth = "EBS",
  ChelaHeights = NULL)), pin = TRUE, pinFile = "tcsam02.pin", minPhase = 1,
  jitter = FALSE, calcTAC = TRUE, HCR = 1), opModInfo = list(path = ".",
  configFile = "OpMod.Configuration.inp", optsFile = "OpMod.Options.inp",
  mpiFile = "OpMod.ParametersInfo,inp", pinFile = "opMod.pin", minPhase =
  5, maxPhase = 6), estModInfo = list(path = ".", configFile =
  "EstMod.Configuration.inp", optsFile = "EstMod.Options.inp", minPhase =
  5), keepFiles = c("tmp.sh", "tcsam02.par"), cleanupAll = FALSE,
  test = TRUE, verbose = TRUE)
}
\arguments{
\item{os}{- 'win', 'mac', 'osx', or 'linux'}

\item{topLevelFolder}{- path to top-level folder for MSE model runs}

\item{model}{- TCSAM02 model executable name}

\item{path2model}{- path to model executable}

\item{HCR}{- integer identifying harvest control rule to use}

\item{numRuns}{- number of complete MSE runs to make}

\item{minRunID}{- integer identifying first run label for the series of runs (default=1)}

\item{firstYr}{- first year for projection (i.e., the assessment year, or max year + 1, of base model)}

\item{numYrs}{- number of years to run model forward}

\item{runBaseModel}{- flag to run base model to calculate TAC and create operating model state for first projection year}

\item{baseModelInfo}{- list of parameters to run base model (if necessary)}

\item{opModInfo}{- list of parameters to run operating model}

\item{estModInfo}{- list of parameters to run estimation model}

\item{keepFiles}{- vector of file names to keep after model run, if cleanupAll is true}

\item{cleanupAll}{- T/F to clean up ALMOST ALL model output files after each run}

\item{test}{- flag (T/F) to run in "test" mode}

\item{verbose}{- flag (T/F) to print diagnostic messages}

\item{configFile}{- path to model configuration file}
}
\value{
- NULL
}
\description{
This functions runs the TCSAM02 MSE for one harvest strategy.
}
\details{
TODO!
}
