% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdRep.CalcABC.R
\name{sdRep.CalcABC}
\alias{sdRep.CalcABC}
\title{Function to calculate ABCs from sd_report output}
\usage{
sdRep.CalcABC(tcsams = NULL, buffer = 0.2, pstar = 0.49, verbose = FALSE)
}
\arguments{
\item{tcsams}{\itemize{
\item a named list of tcsam02.resLst objects
}}

\item{buffer}{\itemize{
\item buffer for ABC as fraction (i.e., 0.20 = 20 \\% yields ABC = 0.8*OFL)
}}

\item{pstar}{\itemize{
\item pstar value for ABC calculation (default = 0.49)
}}

\item{verbose}{flag to print diagnostic info}
}
\value{
dfrABCs - dataframe with MLE management quantities, stdvOFL, ABC.pstar and ABC.buff for each case
}
\description{
This function calculates ABCs from TCSAM02 sd_report output.
}
\details{
Use \code{\link[=sdRep.plotABC]{sdRep.plotABC()}} to produce ABC plots for each case
}
