% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdRep.PlotABC.R
\name{sdRep.PlotABC}
\alias{sdRep.PlotABC}
\title{Plot ABC and OFL from \code{\link[=sdRep.CalcABC]{sdRep.CalcABC()}} output}
\usage{
sdRep.PlotABC(dfrABCs, xlims = NULL)
}
\arguments{
\item{xlims}{\itemize{
\item range for x-axis in plot
}}

\item{dfrABC}{\itemize{
\item dataframe with selected values for ABC.buff, ABC.pstar, and OFL by case
}}
}
\value{
ggplot object
}
\description{
Function to plot ABC and OFL from \code{\link[=sdRep.CalcABC]{sdRep.CalcABC()}} output.
}
\details{
Input dfrABCs should be an output dataframe from \code{\link[=sdRep.CalcABC]{sdRep.CalcABC()}}
}
