% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTCSAM02Params.R
\name{calcTCSAM02Params}
\alias{calcTCSAM02Params}
\title{Calculate parameter values from a TCSAM2013 model for TCSAM02}
\usage{
calcTCSAM02Params(obj, template, out = "Model.ParametersInfo.dat",
  verbose = TRUE)
}
\arguments{
\item{obj}{- tcsam2013.resLst object or tcsam2103.prs object}

\item{template}{- file name of TCSAM02 parameters info template}

\item{out}{- file name for output file}

\item{verbose}{- flag to print debugging info}
}
\value{
character string written to a file
}
\description{
Function to calculate parameter values from a TCSAM2013 model for TCSAM02.
}
\details{
Calculates and exports parameter values for a TCSAM02 model equivalent to those from a TCSAM2013 model run.
}

