% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRep.R
\name{createRep}
\alias{createRep}
\title{Create a tcsam02.rep object from components}
\usage{
createRep(mc = NULL, mr = NULL, model.fits = NULL)
}
\arguments{
\item{mc}{\itemize{
\item model configuration list
}}

\item{mr}{\itemize{
\item model results list
}}

\item{model.fits}{\itemize{
\item model fits list
}}
}
\value{
a tcsam02.rep object
}
\description{
Function to create a tcsam02.rep object from components.
}
\details{
None.
}
