% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createResLst.R
\name{createResLst}
\alias{createResLst}
\title{Create a tcsam02.rep object from components}
\usage{
createResLst(rep = NULL, prs = NULL, std = NULL, ofc = NULL, verbose = FALSE)
}
\arguments{
\item{rep}{\itemize{
\item tcsam02 rep object
}}

\item{prs}{\itemize{
\item tcsam02 prs object
}}

\item{std}{\itemize{
\item tcsam02 std object
}}

\item{ofc}{\itemize{
\item tcsam02 ofc object
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debugging info
}}
}
\value{
a tcsam02.resLst object
}
\description{
Function to create a tcsam02.rep object from components.
}
\details{
None.
}
