% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Data.FleetSizeComps.R
\name{getMDFR.Data.FleetSizeComps}
\alias{getMDFR.Data.FleetSizeComps}
\title{Get fleet size comps by model scenario}
\usage{
getMDFR.Data.FleetSizeComps(
  objs,
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "total", "discard", "retained"),
  data.type = c("input", "aggregated"),
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item single model report list object, or named list of them
}}

\item{fleet.type}{\itemize{
\item fleet type ('fishery' or 'survey')
}}

\item{catch.type}{\itemize{
\item catch type ('index','total','discard', or 'retained')
}}

\item{data.type}{\itemize{
\item "input" or "aggregated"
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\value{
dataframe in canonical format
}
\description{
Function to get fleet size comps by model scenario.
}
\details{
Returned dataframe is in canonical format. "category" will be "n.at.z"
to indicate size compositions (to be consistent across the package).
}
