% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Data.FleetTimeSeries.R
\name{getMDFR.Data.FleetTimeSeries}
\alias{getMDFR.Data.FleetTimeSeries}
\title{Get fleet data components by model scenario}
\usage{
getMDFR.Data.FleetTimeSeries(objs, fleet.type = c("survey", "fishery"),
  data.type = c("abundance", "biomass"), catch.type = c("index",
  "total", "discard", "retained"), ci = 0.8, verbose = FALSE)
}
\arguments{
\item{objs}{- single model report list object, or named list of them}

\item{fleet.type}{- fleet type ('fishery' or 'survey')}

\item{data.type}{- data type ('abundance' or 'biomass')}

\item{catch.type}{- cath type ('index','total','discard', or 'retained')}

\item{ci}{- confidence intervals for time series observations}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
dataframe
}
\description{
Function to get fleet data components by model scenario
}
\details{
Returned dataframe is in canonical format
}
