% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Fits.ChelaHeightData.R
\name{getMDFR.Fits.ChelaHeightData}
\alias{getMDFR.Fits.ChelaHeightData}
\title{Get model fits to chela height data}
\usage{
getMDFR.Fits.ChelaHeightData(objs, verbose = FALSE)
}
\arguments{
\item{objs}{- single model report list object, or named list of them}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
dataframe (or NULL)
}
\description{
Function to get model fits to chela height data.
}
\details{
Returned dataframe is in canonical format
}
