% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Fits.GrowthData.R
\name{getMDFR.Fits.GrowthData}
\alias{getMDFR.Fits.GrowthData}
\title{Get model fits to growth data as a dataframe}
\usage{
getMDFR.Fits.GrowthData(objs, verbose = FALSE)
}
\arguments{
\item{objs}{\itemize{
\item a single tcsam02 rep object, a single tcsam02 resLst object, or a named list of such
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\value{
dataframe (or NULL)
}
\description{
Function to get model fits to growth data as a dataframe.
}
\details{
Returned dataframe is in canonical format.
}
