% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Fits.MaturityOgiveData.R
\name{getMDFR.Fits.MaturityOgiveData}
\alias{getMDFR.Fits.MaturityOgiveData}
\title{Get model fits to maturity ogive data}
\usage{
getMDFR.Fits.MaturityOgiveData(objs, verbose = FALSE)
}
\arguments{
\item{objs}{- single model report list object, or named list of them}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
dataframe (or NULL)
}
\description{
Function to get model fits to maturity ogive data.
}
\details{
Returned dataframe is in canonical format
}
