% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.OFCs.NonDataComponents.R
\name{getMDFR.OFCs.NonDataComponents}
\alias{getMDFR.OFCs.NonDataComponents}
\title{Get non-data components from the objective function for a (list of) TCSAM02 model(s) as a dataframe}
\usage{
getMDFR.OFCs.NonDataComponents(
  obj,
  categories = c("all", "penalties", "penFDevs", "components", "priors"),
  verbose = FALSE
)
}
\arguments{
\item{obj}{\itemize{
\item a single tcsam02 rep object, a single tcsam02 resLst object, or a named list of such
}}

\item{categories}{\itemize{
\item non-data-related objective function components to get ("all","penalties","penFDevs","components","priors")
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\value{
dataframe
}
\description{
Function to get non-data components from the objective function for a (list of) TCSAM02 model(s) as a dataframe.
}
\details{
Returned dataframe has columns
\itemize{
\item{case - model case (blank, to be filled in by caller if obj is not a list of tcsam02.resLst objects)}
\item{category - objective function component category}
\item{type - type name}
\item{element - element name}
\item{level - level name}
\item{wgt - likelihood weight}
\item{nll - negative log-likelihood (unweighted)}
\item{objfun - objective function value}
}
}
