% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.OFLResults.R
\name{getMDFR.OFLResults}
\alias{getMDFR.OFLResults}
\title{Get OFL results as a datframe}
\usage{
getMDFR.OFLResults(objs, verbose = FALSE)
}
\arguments{
\item{objs}{- single tcsam02 resLst object, or named list of them}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
dataframe
}
\description{
Function to get a dataframe of OFL results.
}
\details{
Returned dataframe is NOT in canonical format.
}
