% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.ParametersAtBounds.R
\name{getMDFR.ParametersAtBounds}
\alias{getMDFR.ParametersAtBounds}
\title{Get active parameters estimated at one of their bounds from a set of TCSAM02 model runs as a dataframe}
\usage{
getMDFR.ParametersAtBounds(tcsams, delta = 0.01, verbose = FALSE)
}
\arguments{
\item{tcsams}{\itemize{
\item a tcsam02.resLst object or named list of such
}}

\item{delta}{\itemize{
\item relative fraction of range which defines being at a boundary
}}

\item{verbose}{\itemize{
\item flag to print debugging info
}}
}
\value{
\itemize{
\item a dataframe
}
}
\description{
Function to get active parameters estimated at one of their bounds from a set of TCSAM02 model runs as a dataframe.
}
\details{
Returns a dataframe with active parameters estimated at one of their bounds.
}
