% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.MeanGrowth.R
\name{getMDFR.Pop.MeanGrowth}
\alias{getMDFR.Pop.MeanGrowth}
\title{Get mean growth increments from model results from TCSAM2015 and rsimTCSAM model runs as a dataframe}
\usage{
getMDFR.Pop.MeanGrowth(tcsams, verbose = FALSE)
}
\arguments{
\item{tcsams}{- single TCSAM2015 model report object, or named list of such}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format
}
\description{
Function to get mean growth increments from model results from TCSAM2015 and rsimTCSAM model runs as a dataframe.
}
\details{
Extracts mean growth increments.
}

