% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.NaturalMortality.R
\name{getMDFR.Pop.NaturalMortality}
\alias{getMDFR.Pop.NaturalMortality}
\title{Get natural mortality rates from model results from TCSAM02 model runs as a dataframe}
\usage{
getMDFR.Pop.NaturalMortality(tcsams = NULL, type = "M_c", verbose = FALSE)
}
\arguments{
\item{tcsams}{\itemize{
\item single TCSAM02 model report object, or named list of such
}}

\item{type}{\itemize{
\item flag indicating which M's to extract ('M_cxm', 'M_yxm' or 'M_yxmsz')
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debug info
}}
}
\value{
dataframe in canonical format
}
\description{
Function to get natural mortality rates from model results from TCSAM02 model runs as a dataframe.
}
\details{
Extracts natural mortality rates. Uses \code{reshape2} package.

'type' can be one of:
\itemize{
\item M_c
\item M_cxm
\item M_cxms
\item M_yxm
\item M_yxmsz
\item M_cy
}
}
