% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.prM2M.R
\name{getMDFR.Pop.prM2M}
\alias{getMDFR.Pop.prM2M}
\title{Get molt-to-maturity ogives from model results from TCSAM02 model runs as a dataframe}
\usage{
getMDFR.Pop.prM2M(tcsams, verbose = FALSE)
}
\arguments{
\item{tcsams}{- single tcsam02.rep object, tcsam02.resLst object, or named list of the latter}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format
}
\description{
Function to get molt-to-maturity ogives from model results from TCSAM02 model runs as a dataframe.
}
\details{
Extracts molt-to-maturity ogives.
}

