% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.SdRep.RecAndSSB.R
\name{getMDFR.SdRep.RecAndSSB}
\alias{getMDFR.SdRep.RecAndSSB}
\title{Get sd_report info on recruitment and SSB time series estimates as a dataframe}
\usage{
getMDFR.SdRep.RecAndSSB(tcsams, ci = 0.95, verbose = FALSE)
}
\arguments{
\item{tcsams}{\itemize{
\item a tcsam02.resLst object or named list of such
}}

\item{ci}{\itemize{
\item confidence interval
}}

\item{verbose}{\itemize{
\item flag to print debugging info
}}
}
\value{
a dataframe
}
\description{
Function to get sd_report info on recruitment and SSB time series estimates as a dataframe.
}
\details{
Recruitment values are on ln-scale.
}
