% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Surveys.MaturityOgives.R
\name{getMDFR.Surveys.MaturityOgives}
\alias{getMDFR.Surveys.MaturityOgives}
\title{Get survey maturity ogives by year from model results from TCSAM02 model runs as a dataframe}
\usage{
getMDFR.Surveys.MaturityOgives(tcsams, category = "index", verbose = FALSE)
}
\arguments{
\item{tcsams}{- single tcsam02.rep object, tcsam02.resLst object, or named list of the latter}

\item{category}{- 'index' is only choice}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format
}
\description{
Function to get survey maturity ogives by year from model results from TCSAM02 model runs as a dataframe.
}
\details{
Extracts the estimated survey maturity ogives by year.
}
