% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.ZScores.Abundance.R
\name{getMDFR.ZScores.Abundance}
\alias{getMDFR.ZScores.Abundance}
\title{Get model fits to abundance time series as z-scores for fleet data components}
\usage{
getMDFR.ZScores.Abundance(objs, fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "discarded", "total"),
  verbose = FALSE)
}
\arguments{
\item{objs}{- single model report list object, or named list of them}

\item{fleet.type}{- fleet type ('fishery' or 'survey')}

\item{catch.type}{- catch type ('index','retained','discarded',or 'total')}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
dataframe
}
\description{
Function to get model fits to abundance time series as z-scores for fleet data components.
}
\details{
Uses \code{getMDFR.ZScoresForABData()}.
Returned dataframe is in canonical format.
}

