% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.ZScores.Biomass.R
\name{getMDFR.ZScores.Biomass1}
\alias{getMDFR.ZScores.Biomass1}
\title{Get model fits to biomass time series as z-scores for fleet data components}
\usage{
getMDFR.ZScores.Biomass1(rep, fleet.type, catch.type, verbose)
}
\arguments{
\item{rep}{\itemize{
\item single model report list object
}}

\item{fleet.type}{\itemize{
\item fleet type ('fishery' or 'survey')
}}

\item{catch.type}{\itemize{
\item catch type ('index','retained','discarded',or 'total')
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\value{
dataframe
}
\description{
Function to get model fits to biomass time series as z-scores for fleet data components.
}
\details{
Uses \code{getMDFR.ZScoresForABData()}.
Returned dataframe is in canonical format.
}
