% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.ZScores.PrNatZ.R
\name{getMDFR.ZScores.PrNatZ}
\alias{getMDFR.ZScores.PrNatZ}
\title{Get model fits to size frequencies as z-scores for fleet data components}
\usage{
getMDFR.ZScores.PrNatZ(objs, fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "discarded", "total"),
  residuals.type = c("pearsons", "nlls"), verbose = FALSE)
}
\arguments{
\item{fleet.type}{- fleet type ('fishery' or 'survey')}

\item{catch.type}{- catch type ('index','retained','discarded',or 'total')}

\item{residuals.type}{- residual type ('pearsons' or 'nlls')}

\item{verbose}{- flag (T/F) to print diagnostic info}

\item{repObjs}{- single model report list object, or named list of them}
}
\value{
dataframe
}
\description{
Function to get model fits to size frequencies as z-scores for fleet data components.
}
\details{
Uses \code{getMDFR.ZScores.PrNatZ1()}.
Returned dataframe is in canonical form.
}
