% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOFCs.R
\name{getOFCs}
\alias{getOFCs}
\title{Get model components in the objective function as a tcsam02.ofc (melted dataframe) object}
\usage{
getOFCs(obj, mdl = NULL, verbose = FALSE)
}
\arguments{
\item{obj}{\itemize{
\item a single tcsam02 rep object, a single tcsam02 resLst object, or a named list of such
}}

\item{mdl}{\itemize{
\item name to associate with model results object
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\value{
a melted dataframe with class tcsam02.ofc
}
\description{
Function to get model components in the objective function as a tcsam02.ofc (melted dataframe) object.
}
\details{
If obj is a list of tcsam02 resLst (tcsam.resLst) or rep (tcsam02.rep) objects, then the function
is called recursively for each object, with the associated list component name used as
mdl. If obj is a tcsam02 resLst or rep object and mdl is NULL (the default), then
the configName is used as the model name.

The returned dataframe has columns named
"model", "type", "category", "name", "level", "variable", and "value".

The "variable" column indicates whether the "value" is a weight ('wgt'),
negative log-likelihood ('nll'), or objective function value ('objfun').
}
