% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRep.R
\name{getRep}
\alias{getRep}
\title{Get a TCSAM02 report object (by reading a TCSAM02 model report \link{.rep} file)}
\usage{
getRep(repFile = NULL, verbose = FALSE)
}
\arguments{
\item{repFile}{\itemize{
\item report file from a TCSAM02 model run to source. can be NULL.
}}

\item{verbose}{\itemize{
\item flag (T/F) to prnt diagnostic info
}}
}
\value{
TCSAM02 report object (a list). The returned object will be a list of class 'tcsam02.rep'.
}
\description{
Function to get a TCSAM02 report object
}
\details{
If \code{repFile} is NULL, the user will be prompted to identify a
TCSAM02 model report file from which to source the results object.
The returned object will be a list of class 'tcsam02.rep'.
}
