% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRunCommands.R
\name{getRunCommands}
\alias{getRunCommands}
\title{Generate run commands for a tcsam02 model run}
\usage{
getRunCommands(
  os = "osx",
  model = "tcsam02",
  path2model = "",
  configFile = "",
  pin = FALSE,
  pinFile = NULL,
  mseMode = NULL,
  hess = FALSE,
  minPhase = NULL,
  maxPhase = NULL,
  mcmc = FALSE,
  mc.N = 1e+06,
  mc.save = 1000,
  mc.scale = 1000,
  jitter = FALSE,
  iSeed = NULL,
  calcOFL = FALSE,
  calcTAC = FALSE,
  HCR = 1,
  calcDynB0 = FALSE,
  fullClean = TRUE,
  cleanup = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{os}{\itemize{
\item 'win', 'mac', 'osx', or 'linux'
}}

\item{model}{\itemize{
\item admb model name
}}

\item{path2model}{\itemize{
\item path to model
}}

\item{configFile}{\itemize{
\item filename (including path) to model configuration file
}}

\item{pin}{\itemize{
\item flag (T/F) to use a pin file
}}

\item{pinFile}{\itemize{
\item name of pin file (if different from default)
}}

\item{mseMode}{\itemize{
\item mse mode for model run (NULL \link{off: default}, "opModMode", "estModMode")
}}

\item{hess}{\itemize{
\item flag (T/F) to calculate the hessian
}}

\item{minPhase}{\itemize{
\item start phase (or NULL) for minimization calculations
}}

\item{maxPhase}{\itemize{
\item last phase (or NULL) for minimization calculations
}}

\item{mcmc}{\itemize{
\item flag (T/F) to do mcmc calculations
}}

\item{mc.N}{\itemize{
\item number of mcmc iterations to do
}}

\item{mc.save}{\itemize{
\item number of iterations to skip when saving mcmc calculations
}}

\item{mc.scale}{\itemize{
\item number of iterations to adjust scale for mcmc calculations
}}

\item{jitter}{\itemize{
\item flag (T/F) to use jitter initial values
}}

\item{iSeed}{\itemize{
\item value for random number seed
}}

\item{calcOFL}{\itemize{
\item flag (T/F) to do OFL calculations
}}

\item{calcTAC}{\itemize{
\item flag (T/F) to calculate the TAC for the next fishing year
}}

\item{HCR}{\itemize{
\item integer indicating the Harvest Control Rule used to calculate the TAC
}}

\item{calcDynB0}{\itemize{
\item flag (T/F) to do dynamic B0 calculations
}}

\item{fullClean}{\itemize{
\item flag to clean up almost all files (use when making multiple jitter runs)
}}

\item{cleanup}{\itemize{
\item flag (T/F) to clean up unnecessary files
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\description{
Function to generate a script to make a tcsam02 model run
}
\details{
. If \code{cleanup} is TRUE, then .bar, .b0*, .p0*, .r0*, variance,
EchoOut.dat, CheckFile.dat, and fimn.log files are deleted.\cr
If the path associated with \code{configFile} is a relative one, it should
be relative to the path for model output.
}
