% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.CalcABC.R
\name{mcmc.CalcABC}
\alias{mcmc.CalcABC}
\title{Function to calculate ABC and OFL from an MCMC dataframe}
\usage{
mcmc.CalcABC(mcmc = NULL, dfr = NULL, pstar = 0.49, buffer = 0.2,
  xlims = NULL, title = NULL, verbose = FALSE)
}
\arguments{
\item{mcmc}{- mcmc object}

\item{dfr}{- dataframe with OFL as column (and possibly 'case')}

\item{pstar}{- pstar value for ABC calculation}

\item{buffer}{- buffer for ABC as fraction (i.e., 0.20 = 20 \% yields ABC = 0.8*OFL)}

\item{xlims}{- range for x-axis in plot}

\item{title}{- title for plot}

\item{verbose}{flag to rpint diagnostic info}
}
\value{
dataframe with median results, ABC.pstar and ABC.buffer for each case
}
\description{
This function calculates ABC and OFL from TCSAM02 MCMC output.
}
\details{
uses \code{plotABC} to produce ABC plots for each case
}
