% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.InvertListElement.R
\name{mcmc.InvertListElement}
\alias{mcmc.InvertListElement}
\title{Invert an element of an mcmc list}
\usage{
mcmc.InvertListElement(mcmc, element = "", verbose = FALSE)
}
\arguments{
\item{mcmc}{- mcmc list}

\item{element}{- name of element to extract}

\item{verbose}{- T/F to print diagnostic info}
}
\value{
a list
}
\description{
Function to invert an element of an mcmc list
}
\details{
None
}
