% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMSE.R
\name{runMSE}
\alias{runMSE}
\title{Function to run the TCSAM02 MSE for one harvest strategy}
\usage{
runMSE(
  os = "osx",
  topLevelFolder = ".",
  model = "tcsam02",
  path2model = "",
  HCR = 1,
  numRuns = 3,
  minRunID = 1,
  firstYr = 2018,
  numYrs = 10,
  iSeeds = NULL,
  runBaseModel = TRUE,
  baseModelInfo = list(path = ".", configFile = "ModelConfig.inp", datasets = list(file
    = "Datasets.inp", components = list(Bio = "Info", Fishery = c("TCF", "SCF", "GTF",
    "RKF"), Survey = "NMFS", Growth = "EBS", ChelaHeights = NULL)), pin = TRUE, pinFile =
    "tcsam02.pin", minPhase = 1, jitter = FALSE, iSeed = NULL, calcTAC = TRUE, HCR = 1),
  opModInfo = list(path = ".", configFile = "OpMod.Configuration.inp", optsFile =
    "OpMod.Options.inp", mpiFile = "OpMod.ParametersInfo,inp", pinFile = "opMod.pin",
    minPhase = 5, maxPhase = 6),
  estModInfo = list(path = ".", configFile = "EstMod.Configuration.inp", optsFile =
    "EstMod.Options.inp", minPhase = 5),
  keepFiles = c("tmp.sh", "tcsam02.par"),
  cleanupAll = FALSE,
  test = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{os}{\itemize{
\item 'win', 'mac', 'osx', or 'linux'
}}

\item{topLevelFolder}{\itemize{
\item path to top-level folder for MSE model runs
}}

\item{model}{\itemize{
\item TCSAM02 model executable name
}}

\item{path2model}{\itemize{
\item path to model executable
}}

\item{HCR}{\itemize{
\item integer identifying harvest control rule to use
}}

\item{numRuns}{\itemize{
\item number of complete MSE runs to make
}}

\item{minRunID}{\itemize{
\item integer identifying first run label for the series of runs (default=1)
}}

\item{firstYr}{\itemize{
\item first year for projection (i.e., the assessment year, or max year + 1, of base model)
}}

\item{numYrs}{\itemize{
\item number of years to run model forward
}}

\item{iSeeds}{\itemize{
\item matrix \link{(number of total possible runs) x (number of years)} of random number seed values (or NULL)
}}

\item{runBaseModel}{\itemize{
\item flag to run base model to calculate TAC and create operating model state for first projection year
}}

\item{baseModelInfo}{\itemize{
\item list of parameters to run base model (if necessary)
}}

\item{opModInfo}{\itemize{
\item list of parameters to run operating model
}}

\item{estModInfo}{\itemize{
\item list of parameters to run estimation model
}}

\item{keepFiles}{\itemize{
\item vector of file names to keep after model run, if cleanupAll is true
}}

\item{cleanupAll}{\itemize{
\item T/F to clean up ALMOST ALL model output files after each run
}}

\item{test}{\itemize{
\item flag (T/F) to run in "test" mode
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic messages
}}

\item{configFile}{\itemize{
\item path to model configuration file
}}
}
\value{
\itemize{
\item NULL
}
}
\description{
This functions runs the TCSAM02 MSE for one harvest strategy.
}
\details{
\code{numRuns} is the number of \emph{successful} iterations the MSE is to complete before stopping.
If a model run fails for some reason during an iteration in the MSE, the iteration is restarted from the
\code{firstYr} with a \emph{new} seed for the random number generator (RNG). If the input matrix \code{iSeeds}
is NULL, then the new RNG seed is based on the system clock. However, if the matrix is supplied,
the seed is taken from the first element in the "next" row of the matrix (a counter for the
\emph{total} number of MSE iterations is used to track this; a separate counter tracks the number of
\emph{successful} iterations). Thus, if \code{iSeeds} is supplied, the number of rows in the matrix needs to be
larger than (or equal to) the \emph{total} number of MSE iterations actually undergone to achieve \code{numRuns}
\emph{successful} iterations.
}
