% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTCSAM02.R
\name{runTCSAM02}
\alias{runTCSAM02}
\title{Function to run TCSAM02.}
\usage{
runTCSAM02(os = "osx", path = ".", model = "tcsam02",
  path2model = "", configFile = "", pin = FALSE, pinFile = NULL,
  mseMode = NULL, minPhase = 1, maxPhase = NULL, calcOFL = FALSE,
  calcTAC = FALSE, HCR = 1, calcDynB0 = FALSE, hess = FALSE,
  mcmc = FALSE, mc.N = 1e+06, mc.save = 1000, mc.scale = 1000,
  jitter = FALSE, jit.seed = NULL, saveResults = hess, test = TRUE,
  cleanup = TRUE, verbose = FALSE)
}
\arguments{
\item{os}{- 'win', 'mac', 'osx', or 'linux'}

\item{path}{- path for model output}

\item{model}{- TCSAM02 model executable name}

\item{path2model}{- path to model executable}

\item{configFile}{- filename (including path) to model configuration file}

\item{pin}{- T/F to use a pin file}

\item{minPhase}{- phase to start minimization (or NULL)}

\item{maxPhase}{- final minimization phase (or NULL)}

\item{calcOFL}{- flag (T/F) to perform OFL calculations}

\item{calcTAC}{- flag (T/F) to calculate the TAC for the next fishing year}

\item{HCR}{- integer indicating the Harvest Control Rule used to calculate the TAC}

\item{calcDynB0}{- flag to calculate dynamic B0}

\item{hess}{- T/F to compute hessian (and .std file)}

\item{mcmc}{- T/F to run mcmc}

\item{mc.N}{- number of mcmc iterations to do}

\item{mc.save}{- number of iterations to skip when saving mcmc calculations}

\item{mc.scale}{- number of iterations to adjust scale for mcmc calculations}

\item{jitter}{- T/F to jitter parameters}

\item{jit.seed}{- seed for random number generator (or NULL)}

\item{saveResults}{- T/F to save results to ModelResults.RData as a tcsam02.resLst object using \code{getResLst(...)}}

\item{test}{- flag (T/F) to run function in "test" mode}

\item{cleanup}{- flag (T/F) to clean up some output files}

\item{verbose}{- flag to print debugging info}
}
\value{
- dataframe of class 'tcam02.par', with 2 columns (name, value) with jitter jit.seed (if jittered)
and par file info, or NULL if par file does not exist.
}
\description{
This function runs a TCSAM02 model once.
}
\details{
This function creates a shell script ('./tmp.sh') in the
working directory and uses it to run a version of the TCSAM02 model.\cr
Initial model parameters can be jittered based on the system clock time as a jit.seed
to the random number generator. The jit.seed and final objective function value are
saved for each model run in a csv file (the value of out.csv).


If the path associated with \code{configFile} is a relative one, it should
be relative to the \code{path} variable. If saveResults=TRUE, getResLSt() is used to read in
the report file, prs file, and std files are read in and the resulting tcsam02.resLst object is
saved to 'ModelResults.RData'. If jitter=TRUE, hess=FALSE, and cleanup=TRUE, then most output files
(including the .rep files) are deleted after the model run to save disk space.
}
